REFACTOR: change name User to Player
